# Speed up tga decoding and add new benchmark comparison