--- a/Directory.Build.targets
+++ b/Directory.Build.targets
@@ -22,18 +22,19 @@
 
   <!-- Package versions for package references across all projects -->
   <ItemGroup>
-    <PackageReference Update="BenchmarkDotNet" Version="0.11.5" />
-    <PackageReference Update="Colourful" Version="2.0.2" />
+    <PackageReference Update="BenchmarkDotNet" Version="0.12.0" />
+    <PackageReference Update="Colourful" Version="2.0.3" />
     <PackageReference Update="Magick.NET-Q16-AnyCPU" Version="7.14.4" />
     <PackageReference Update="Microsoft.Net.Compilers.Toolset" Version="3.3.1" />
     <PackageReference Update="Microsoft.NET.Test.Sdk" Version="15.9.0" />
     <PackageReference Update="Moq" Version="4.10.0" />
+    <PackageReference Update="Pfim" Version="0.9.1" />
     <PackageReference Update="SixLabors.Core" Version="1.0.0-beta0008" />
     <PackageReference Update="SixLabors.Fonts" Version="1.0.0-beta0009" />
     <PackageReference Update="SixLabors.Shapes" Version="1.0.0-beta0009" />
     <PackageReference Update="SixLabors.Shapes.Text" Version="1.0.0-beta0009" />
     <PackageReference Update="StyleCop.Analyzers" Version="1.1.118" />
-    <PackageReference Update="System.Drawing.Common" Version="4.5.1" />
+    <PackageReference Update="System.Drawing.Common" Version="4.7.0" />
     <PackageReference Update="System.IO.Compression" Version="4.3.0" />
     <PackageReference Update="System.IO.UnmanagedMemoryStream" Version="4.3.0" />
     <PackageReference Update="System.Runtime.CompilerServices.Unsafe" Version="4.5.1" />