--- a/tests/ImageSharp.Benchmarks/Config.cs
+++ b/tests/ImageSharp.Benchmarks/Config.cs
@@ -1,8 +1,9 @@
-﻿// Copyright (c) Six Labors and contributors.
+// Copyright (c) Six Labors and contributors.
 // Licensed under the Apache License, Version 2.0.
 
 using BenchmarkDotNet.Configs;
 using BenchmarkDotNet.Diagnosers;
+using BenchmarkDotNet.Environments;
 using BenchmarkDotNet.Jobs;
 
 namespace SixLabors.ImageSharp.Benchmarks
@@ -19,8 +20,8 @@ namespace SixLabors.ImageSharp.Benchmarks
             public ShortClr()
             {
                 this.Add(
-                    Job.Clr.WithLaunchCount(1).WithWarmupCount(3).WithIterationCount(3),
-                    Job.Core.WithLaunchCount(1).WithWarmupCount(3).WithIterationCount(3)
+                    Job.Default.With(ClrRuntime.Net472).WithLaunchCount(1).WithWarmupCount(3).WithIterationCount(3),
+                    Job.Default.With(CoreRuntime.Core21).WithLaunchCount(1).WithWarmupCount(3).WithIterationCount(3)
                 );
             }
         }