### Prerequisites

- [x] I have written a descriptive pull-request title
- [x] I have verified that there are no overlapping [pull-requests](https://github.com/SixLabors/ImageSharp/pulls) open
- [x] I have verified that I am following matches the existing coding patterns and practice as demonstrated in the repository. These follow strict Stylecop rules :cop:.
- [x] I have provided test coverage for my change (where applicable)

### Description
Couldn't help myself when I discovered a new best-in-class decoder. 

**Before** (Both benchmarks made using `Bgr24` after update to benchmark)

|            Method |       Runtime |           TestImage |         Mean |        Error |      StdDev | Ratio |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------ |-------------- |-------------------- |-------------:|-------------:|------------:|------:|-------:|------:|------:|----------:|
| 'ImageMagick Tga' |    .NET 4.7.2 | Tga/targa_24bit.tga | 2,244.805 us | 3,012.486 us | 165.1245 us | 1.000 |      - |     - |     - |   13680 B |
|  'ImageSharp Tga' |    .NET 4.7.2 | Tga/targa_24bit.tga |    98.776 us |    45.871 us |   2.5144 us | 0.044 | 0.2441 |     - |     - |    1260 B |
|        'Pfim Tga' |    .NET 4.7.2 | Tga/targa_24bit.tga |     8.045 us |    18.927 us |   1.0375 us | 0.004 | 0.0610 |     - |     - |     313 B |
|                   |               |                     |              |              |             |       |        |       |       |           |
| 'ImageMagick Tga' | .NET Core 2.1 | Tga/targa_24bit.tga | 2,368.527 us | 2,246.905 us | 123.1604 us | 1.000 |      - |     - |     - |   13307 B |
|  'ImageSharp Tga' | .NET Core 2.1 | Tga/targa_24bit.tga |   119.434 us |    81.333 us |   4.4582 us | 0.050 | 0.2441 |     - |     - |    1200 B |
|        'Pfim Tga' | .NET Core 2.1 | Tga/targa_24bit.tga |     6.229 us |     3.035 us |   0.1664 us | 0.003 | 0.0610 |     - |     - |     280 B |

**After**

|            Method |       Runtime |           TestImage |         Mean |        Error |     StdDev | Ratio |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------ |-------------- |-------------------- |-------------:|-------------:|-----------:|------:|-------:|------:|------:|----------:|
| 'ImageMagick Tga' |    .NET 4.7.2 | Tga/targa_24bit.tga | 1,778.965 us | 1,711.088 us | 93.7905 us | 1.000 | 1.9531 |     - |     - |   13668 B |
|  'ImageSharp Tga' |    .NET 4.7.2 | Tga/targa_24bit.tga |    38.659 us |     6.886 us |  0.3774 us | 0.022 | 0.3052 |     - |     - |    1316 B |
|        'Pfim Tga' |    .NET 4.7.2 | Tga/targa_24bit.tga |     6.752 us |    10.268 us |  0.5628 us | 0.004 | 0.0687 |     - |     - |     313 B |
|                   |               |                     |              |              |            |       |        |       |       |           |
| 'ImageMagick Tga' | .NET Core 2.1 | Tga/targa_24bit.tga | 1,407.585 us |   124.215 us |  6.8087 us | 1.000 | 1.9531 |     - |     - |   13307 B |
|  'ImageSharp Tga' | .NET Core 2.1 | Tga/targa_24bit.tga |    17.958 us |     9.352 us |  0.5126 us | 0.013 | 0.2747 |     - |     - |    1256 B |
|        'Pfim Tga' | .NET Core 2.1 | Tga/targa_24bit.tga |     5.645 us |     2.279 us |  0.1249 us | 0.004 | 0.0610 |     - |     - |     280 B |

<!-- Thanks for contributing to ImageSharp! -->