@brianpopow 2020-01-07
@JimBobSquarePants that's some nice speedup. I need to look into Pfim source code myself.